feat: voice conversation mode, SSRF guard, RSI self-critique, and knowledge schema
Voice subsystem:
- Add conversation mode with WebSocket session, TTS/streaming-TTS, and playback
- Add auth endpoint for voice service authentication
- Harden speaker streaming path with gate on streaming (no more batch-only fallback)
- Extend recorder with stream_record improvements and discard_recording coverage
- Add voice config expansions for conversation and TTS backends

Security / sandbox:
- Add SSRF egress guard (crates/jfc-engine/src/tools/ssrf.rs)
- Harden sandbox egress rules and approval flows
- Extend permissions with finer-grained tool approval

Engine / tools:
- Add daemon tool helpers and subagent tool wrappers
- Expand plan_recall with structured recall support
- Add continuation and queue drain fixes
- Extend events, session core, and stream request context

RSI / learning:
- Add prompt_miner and self_critique modules under jfc-learn
- Add examples: prompt_skill_mine, prune_noise, recall_improve, rsi_foundation, self_critique_backfill

Knowledge / memory:
- Expand jfc-knowledge with full lib, schema, and definitions
- Add session_mine for mining session history into knowledge graph
- Improve memory recall scoring

MCP / config:
- Extend MCP registry with richer tool advertising
- Add claude_settings expansions and config lib updates

Markdown:
- Add diagram rendering support (jfc-markdown)

Research artifacts:
- Add axon-papers corpus (PDFs + extracted text + synthesis notes)
- Add codegraph_live_viz SPEC and voice_speaker_identity SPEC

+++ crates/jfc/src/render/messages.rs
@@ -1295,7 +1295,7 @@ fn working_activity_subject(app: &App) -> String {
         return format!("dispatching {count} tool call{}", plural_suffix(count));
     }
     if app.engine.in_flight_tool_batches > 0 {
-        let count = app.engine.in_flight_tool_batches as usize;
+        let count = app.engine.in_flight_tool_batches;
         return format!("running {count} tool batch{}", plural_suffix(count));
     }
     if !app.engine.in_progress_tool_use_ids.is_empty() {
@@ -1303,7 +1303,7 @@ fn working_activity_subject(app: &App) -> String {
         return format!("waiting on {count} tool result{}", plural_suffix(count));
     }
     if app.engine.in_flight_eager_dispatches > 0 {
-        let count = app.engine.in_flight_eager_dispatches as usize;
+        let count = app.engine.in_flight_eager_dispatches;
         return format!("dispatching {count} eager tool{}", plural_suffix(count));
     }
     let alive_agents = app
